Refactor alerts
