+++ Applite/Utilities/Brew Installation/DependencyError.swift
@@ -7,15 +7,20 @@
 
 import Foundation
 
-enum DependencyError: Error {
+enum DependencyError: LocalizedError {
     case xcodeCommandLineToolsTimeout
-}
 
-extension DependencyError: LocalizedError {
-    public var errorDescription: String? {
+    var errorDescription: String? {
         switch self {
         case .xcodeCommandLineToolsTimeout:
-            return "Xcode Command Line Tools install timeout"
+            return "Couldn't install Xcode Command Line Tools"
+        }
+    }
+
+    var failureReason: String? {
+        switch self {
+        case .xcodeCommandLineToolsTimeout:
+            return "Couldn't install Xcode Command Line Tools in a reasonable amount of time"
         }
     }
 }
