+++ Applite/Views/App Views/App View/AppView+UninstallButton.swift
@@ -1,0 +1,31 @@
+//
+//  AppView+UninstallButton.swift
+//  Applite
+//
+//  Created by Milán Várady on 2024.12.26.
+//
+
+import SwiftUI
+
+extension AppView {
+    struct UninstallButton: View {
+        @StateObject var cask: Cask
+        @EnvironmentObject var caskData: CaskData
+
+        @State var showingError = false
+
+        var body: some View {
+            Button {
+                Task {
+                    await MainActor.run { cask.progressState = .busy(withTask: "Uninstalling") }
+
+                    _ = await cask.uninstall(caskData: caskData)
+                }
+            } label: {
+                Image(systemName: "trash.fill")
+                    .font(.system(size: 20))
+            }
+            .foregroundColor(.primary)
+        }
+    }
+}
